Upgrade: Improved ownership to allow for smoother interactions.
PUN allows you to takeover control of an object, which was the requested behavior before, but streamlined. To keep it so that we can restrict access to manipulating an object but without introducing the clunkier object ownership requests, the ownership is assumed to be free to grab whenever. If you want to create a "locked" object, you can call InstantiateOwnedObject instead of Instantiate in the ObjectInteractionManager. Further, or more in-depth, manipulation has to be accessed through the NetworkManager, since you're working more with synchronization and object ownership which is application or context sensitive.